move Network/Socket to Internal
